Extract sensor rotation screen building blocks
